Serialize TriggerTypeClass field by field
TriggerTypeClass, its two TEventClass and two TActionClass members and
the AbstractTypeClass base part move to Serialize(). The event and
action unions travel through their widest member, which for TEventClass
becomes int32_t in place of long. NoInit constructors, never-defined
Load/Save and the empty Code/Decode_Pointers go; members get default
initializers. The existing default constructor already did nothing but
initialize, so it doubles as the shell for TFixedIHeapClass::Load.

+++ src/ra/tevent.cc
@@ -442,7 +442,7 @@ bool TEventClass::operator()(TDEventClass& td, TEventType event,
       *destroyed.
       */
       case TEVENT_NBUILDINGS_DESTROYED:
-        if (hptr->BuildingsLost < Data.Value) {
+        if (static_cast<int>(hptr->BuildingsLost) < Data.Value) {
           return false;
         }
         break;
@@ -451,7 +451,7 @@ bool TEventClass::operator()(TDEventClass& td, TEventType event,
       **	Verify that the specified number of units have been destroyed.
       */
       case TEVENT_NUNITS_DESTROYED:
-        if (hptr->UnitsLost < Data.Value) {
+        if (static_cast<int>(hptr->UnitsLost) < Data.Value) {
           return false;
         }
         break;
